Fix wrong search query in subject
If we use a standard port and try to search it for a string with mandatory port specification (function getHostWithPort()), the string will not be found. We need to either look for getUrl() in the subject or change the subject from `request()->getUrl()` to `request()->getScheme() . '://' . request()->getHostWithPort()` because the getUrl() function does not output the port if it is standard.